Moved custom tab name manager to the custom hook from the panel
